backend/requirements.txt
# CompareIQ AI — Backend Dependencies
# Author: Raja Abimanyu N | Data Scientist | AI & Applied Machine Learning

# ── Core ────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30
aiosqlite>=0.20.0          # SQLite async driver (zero-config local dev)
asyncpg>=0.29.0            # Postgres async driver (production)
alembic>=1.13.0

# ── Redis ────────────────────────────────────────────────────────────────────
redis>=5.0.4

# ── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1              # Pin to 4.0.1 — passlib 1.7.x is incompatible with bcrypt 4.1+
python-multipart>=0.0.9

# ── AI / LLM ─────────────────────────────────────────────────────────────────
openai>=1.35.0
langchain>=0.2.5
langchain-openai>=0.1.14
langchain-groq>=0.1.6
langgraph>=0.1.14
langchain-core>=0.2.10
langchain-community>=0.2.6

# ── Vector DB ────────────────────────────────────────────────────────────────
qdrant-client>=1.9.0

# ── Parsing ──────────────────────────────────────────────────────────────────
pymupdf>=1.24.5            # PDF parsing  (pip name: PyMuPDF)
beautifulsoup4>=4.12.3     # HTML parsing
httpx>=0.27.0              # Async HTTP client
lxml>=5.2.0                # Fast HTML parser

# ── Export ───────────────────────────────────────────────────────────────────
reportlab>=4.2.0           # PDF report generation

# ── Background tasks ─────────────────────────────────────────────────────────
celery>=5.4.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
